finished tests - added some logging

+++ src/main/java/it/ldas/imageprocessor/model/entity/ImageStatistics.java
@@ -20,11 +20,11 @@ import java.math.BigDecimal;
 @Synchronize("images")
 public class ImageStatistics {
 
-    @Id  // id per hibernate
+    @Id
     private Long id;
     private Integer totalImages;
-    private BigDecimal totalSize;
-    private BigDecimal totalPixels;
+    private Long totalSize;
+    private Long totalPixels;
     private Integer receivedImages;
     private Integer processingImages;
     private Integer processedImages;
